Extract `EncoderOptions` into a standalone module and integrate it into `TokenEncoderBuilder` for configuration management.
